requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
sqlparse==0.5.0
tzdata==2024.1
